Webapp/requirements.txt
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1
openpyxl==3.1.5
Pillow==10.4.0
